Add logic to only handle signal events for default cgroup hierarchy
Since with cgroupV1 there are multiple different cgroups involved,
kvisor should only handle cgroup signal events for a certain cgroup
hierarchy. In our case we always use cpusets as our default hierarchy,
so we only handle cgroup mkdir/rmdir events for this hierarchy.

All of this only applies to cgroupsV1 as with V2 cgroups are unified
under a single root.