Report only lower 32-bits of cgroup id from eBPF
With cgroupV1, the cgroup id contains the thread id in the upper 32bits.
Kvisor is not really concerned with cgroups on this level, hence the
logic has been changed to only report the lower 32bits to userspace.

For cgroupsV2 nothing changes, as it anyway should be contained in the
lower 32bits.